pyproject.toml
[tool.poetry]
name = "say-botel"
version = "0.1.0"
description = "This serviece provides the telegram related stuff/behavior to other SayIf services"
license = "LGPL-3.0-or-later"
repository = "https://github.com/SayIfOrg/say_botel"
authors = ["Amir Khalife <eng.amir.bu@gmail.com>"]
packages = [{include = "botel"},{include = "main.py"}]

    [tool.poetry.dependencies]
    python = "~3.9 || ~3.10 || ~3.11"
    pyTelegramBotAPI = "~4.7.0"
    aiohttp = "*" # required by "pyTelegramBotAPI" async
    alembic = "~1.8.1"
    asyncpg = "~0.27"  # alternatively you can use psycopg3
    beautifulsoup4 = "~4.11"
    grpcio = "~1.50"
    grpcio-tools = "~1.50"
    sqlalchemy = "~1.4.40"
    say-protos = {git = "https://github.com/SayIfOrg/say_protos.git", subdirectory = "packages/python", rev = "cd7f74268a6"}
    gql = "~3.4.1"

    [tool.poetry.group.dev.dependencies]
    python-dotenv = "~1.0"

    [tool.poetry.group.local.dependencies]
    # local dependencies to develop in sync with dependent packages
    # change the relative path to fit your usecagse

#     say-protos = {path = "../path-to-local-dependency/say_protos/packages/python", develop = true}

    [tool.poetry.extras]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
extend_skip="alembic"
lines_after_imports=2
profile = "black"
py_version=39
